// UI/src/app/model/raw-material-stock.ts
export class RawMaterialStock {
    stockId: String;
    orderId: String;
    name: String;
    price_per_unit: number;
    quantityValue: number;
    quantityUnit: number;
    price: number;
    warehouseId: String;
    deliveryDate: Date;
    manufacturingDate: Date;
    expiryDate: Date;
    qualityCheck: String;
    processDate: Date;
}